feat: rework PasswordRecovery screen to match the pattern MVVM
