+++ app/src/main/res/layout/activity_password_recovery.xml
@@ -4,7 +4,27 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PasswordRecoveryActivity">
+    tools:context=".screens.screenPasswordRecovery.PasswordRecoveryActivity">
+
+    <Button
+        android:id="@+id/buttonGoBack3"
+        android:layout_width="50dp"
+        android:layout_height="50dp"
+        android:layout_margin="5dp"
+        android:layout_weight="0"
+        android:background="#00FFFFFF"
+        android:onClick="onClickGoBack"
+        app:icon="@drawable/ic_baseline_arrow_back_24"
+        app:iconGravity="textStart"
+        app:iconSize="34sp"
+        app:iconTint="#818181"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.0"
+        tools:ignore="SpeakableTextPresentCheck" />
 
     <EditText
         android:id="@+id/editTextEmailAddressRec"
@@ -38,12 +58,16 @@
         android:id="@+id/button"
         android:layout_width="340dp"
         android:layout_height="60dp"
+        android:backgroundTint="#FFFFFF"
         android:onClick="onClickSendRecoveryMail"
         android:text="@string/passwordrecovery_send_mail"
+        android:textColor="#1FA810"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/editTextEmailAddressRec"
-        app:layout_constraintVertical_bias="0.954" />
+        app:layout_constraintVertical_bias="0.954"
+        app:strokeColor="#A1A1A1"
+        app:strokeWidth="0.1sp" />
 </androidx.constraintlayout.widget.ConstraintLayout>
